# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is backend api service for tks platform",
        "title": "tks-api service",
        "contact": {
            "name": "taekyu.kang@sk.com",
            "email": "taekyu.kang@sk.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "tks-api-dev.taco-cat.xyz:9110",
    "basePath": "/api/1.0/",
    "paths": {
        "/app-groups": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get appGroup list by giving params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Get appGroup list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "clusterId",
                        "name": "clusterId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAppGroupsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Install appGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Install appGroup",
                "parameters": [
                    {
                        "description": "create appgroup request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAppGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAppGroupResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Uninstall appGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Uninstall appGroup",
                "parameters": [
                    {
                        "description": "body",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/app-groups/{appGroupId}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get appGroup detail by appGroupId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Get appGroup detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "appGroupId",
                        "name": "appGroupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAppGroupResponse"
                        }
                    }
                }
            }
        },
        "/app-groups/{appGroupId}/applications": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get applications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Get applications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "appGroupId",
                        "name": "appGroupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "applicationType",
                        "name": "applicationType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetApplicationsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppGroups"
                ],
                "summary": "Update application",
                "parameters": [
                    {
                        "description": "body",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/app-serve-apps/{appServeAppId}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get appServeApp by giving params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppServeApps"
                ],
                "summary": "Get appServeApp",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AppServeApp"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user detail",
                        "schema": {
                            "$ref": "#/definitions/domain.LoginResponse"
                        }
                    }
                }
            }
        },
...
        "/cloud-settings": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get CloudSettings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CloudSettings"
                ],
                "summary": "Get CloudSettings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "show all organizations",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetCloudSettingsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create CloudSetting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CloudSettings"
                ],
                "summary": "Create CloudSetting",
                "parameters": [
                    {
                        "description": "create cloud setting request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateCloudSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateCloudSettingResponse"
                        }
                    }
                }
            }
        },
...
        "/histories": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get histories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histories"
                ],
                "summary": "Get histories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.History"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateAppGroupRequest": {
            "type": "object",
            "required": [
                "clusterId",
                "name"
            ],
            "properties": {
                "appGroupType": {
                    "type": "string",
                    "enum": [
                        "LMA",
                        "SERVICE_MESH"
                    ]
                },
                "clusterId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.CreateAppGroupResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "domain.GetAppGroupResponse": {
            "type": "object",
            "properties": {
                "appGroup": {
                    "$ref": "#/definitions/domain.AppGroupResponse"
                }
            }
        },
        "domain.GetAppGroupsResponse": {
            "type": "object",
            "properties": {
                "appGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AppGroupResponse"
                    }
                }
            }
        },
        "domain.GetApplicationsResponse": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ApplicationResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}